+++ supabase/functions/get-quiz-attempt/index.ts
@@ -1,71 +1,16 @@
-import "jsr:@supabase/functions-js/edge-runtime.d.ts"
-import { OptionsRsp, SuccessResponse } from "../_shared/helpers.ts";
-import { getUserById, verifyAdministrator } from "../_shared/auth.ts";
-import { getRows } from "../_shared/database.ts";
+import "jsr:@supabase/functions-js/edge-runtime.d.ts";
+import { z } from "npm:zod";
+import EdgeFunctionRequest, { RunParams } from "../_shared/EdgeFunctionRequest.ts";
+import getQuizAttempt from "./getQuizAttempt.ts";
 
-admin only!!!
-
-Deno.serve(async (req) => {
-
-    if (req.method === 'OPTIONS') {
-        return OptionsRsp();
-    }
-
-    const adminStatus = await verifyAdministrator(req);
-    if (adminStatus instanceof Response) return adminStatus;
-
-    const { quizAttemptId } = await req.json();
-
-    const quizAttemptQuery = await getRows({ table: 'quiz_attempt', conditions: ['eq', 'id', quizAttemptId] });
-    if (quizAttemptQuery instanceof Response) return quizAttemptQuery;
-    const quizAttempt = quizAttemptQuery[0];
-
-    const courseQuery = await getRows({ table: 'course', conditions: ['eq', 'id', quizAttempt.course_id] });
-    if (courseQuery instanceof Response) return courseQuery;
-    const course = courseQuery[0];
-
-    const user = await getUserById(req, quizAttempt.user_id);
-    if (user instanceof Response) return user;
-
-    const questionAttempts = await getRows({ table: 'quiz_question_attempt', conditions: ['eq', 'quiz_attempt_id', quizAttemptId] });
-    if (questionAttempts instanceof Response) return questionAttempts;
-
-    const questions = await getRows({ table: 'quiz_question', conditions: ['eq', 'course_id', quizAttempt.course_id] });
-    if (questions instanceof Response) return questions;
-
-    const saQuestions = questionAttempts.filter((q) => q.type === "SA").map((q) => {
-        const question = questions.find((question) => question.id === q.quiz_question_id);
-        return {
-            question: question.question,
-            marks: question.marks,
-            questionAttemptId: q.id,
-            response: q.response,
-            marksAchieved: q.marks_achieved
-        };
-    });
-
-    const otherQuestions = questionAttempts.filter((q) => q.type !== "SA").map((q) => {
-        const question = questions.find((question) => question.id === q.quiz_question_id);
-        return {
-            question: question.question,
-            type: question.type,
-            answers: question.answers,
-            correctAnswer: question.correct_answer,
-            marks: question.marks,
-            response: q.response,
-            marksAchieved: q.marks_achieved
-        };
-    });
-
-    const response = {
-        courseName: course.name,
-        submitterName: user.user_metadata.name,
-        completionTime: new Date(quizAttempt.end_time),
-        saQuestions: saQuestions,
-        otherQuestions: otherQuestions,
-        score: quizAttempt.score,
-        markingInfo: quizAttempt.markerInfo
+Deno.serve(async (req: Request) => {
+    const parameters: RunParams = {
+        metaUrl: import.meta.url,
+        req: req,
+        schemaRecord: { quizAttemptId: z.string() },
+        endpointFunction: getQuizAttempt,
+        adminOnly: true
     };
 
-    return SuccessResponse(response);
+    return await EdgeFunctionRequest.run(parameters);
 });
